[ci] Create secrets_test.go
This commit moves existing tests revolving around secrets into a new
test file secrets_test.go
